show the possibility of deleting data in a dialog
The same dialog will show both in the DNP list and in the DNP dedicated page. I will show the following text:
This action cannot be undone. If you do NOT want to keep ${name}'s data remove it permanently with the "Remove DNP + data" option.
Then it will give three options
- Cancel
- Remove
- Remove DNP + data